pyproject.toml
[project]
name = "phlyght"
license = {file = "COPYING"}
description = "An async Python library for controlling Philips Hue lights."
authors = [{name = "", email = "ra@tcp.direct"}]
keywords = ["philips", "hue", "lights", "async", "asyncio", "python3"]
dependencies = [
    "httpx>=0.23.1",
    "pydantic>=1.10.0",
    "yarl>=1.8.0",
    "ujson>=5.0.0",
    "rich>=12.6.0",
]
version = "1.0.0"

[tool.setuptools.packages.find]
include = ["phlyght"]

[tool.poetry]
name = "lights"
version = "1.0.1"
description = "An async Python library for controlling Philips Hue lights."
authors = ["Ra <ra@tcp.direct>"]

[tool.poetry.dependencies]
python = ">=3.11,<4.0.0"
httpx = ">=0.23.1"
pydantic = ">=1.10.0"
yarl = ">=1.8.0"
ujson = ">=5.6.0"
rich = ">=12.6.0"
aiofiles = ">=22.1.0"
pyyaml = "^6.0"
loguru = "^0.6.0"
orjson = "^3.8.5"

[tool.poetry.dev-dependencies]
black = ">=22.10.0"

[tool.poetry.group.dev.dependencies]
pycodestyle = "^2.10.0"
pylint = "^2.15.7"
mypy = "^0.991"
flake8 = "^6.0.0"

[tool.poetry.group.linux.dependencies]
uvloop = "^0.17.0"

[tool.poetry.group.linux]
optional = true


[build-system]
requires = ["setuptools >= 39.2.0", "wheel", "poetry-core>=1.0.0"]
build-backend = "setuptools.build_meta"

[tool.flake8]
ignore = ["W503"]
extras = ["E501", "E203"]

[tool.pyright]
pythonVersion = "3.11.1"
pythonPlatform = "Linux"
include = [ "*.py" ]
ignore = ["reportGeneralTypeIssues"]
reportMissingImports = true
reportMissingTypeStubs = false
